% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formated_output.R
\name{add_empty_rows_between_groups}
\alias{add_empty_rows_between_groups}
\title{Insert Empty Rows into a Data Frame}
\usage{
add_empty_rows_between_groups(df, col)
}
\arguments{
\item{df}{The data frame.}

\item{col}{The column used to identify groups.}
}
\value{
A modified dataset with empty rows separating distinct groups.
}
\description{
Internal helper function to insert empty rows between grouped elements.
}
\keyword{internal}
