% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_responses.R
\name{add_to_cleaning_log_other_recode}
\alias{add_to_cleaning_log_other_recode}
\title{Add Recoding of Other Response to Cleaning Log}
\usage{
add_to_cleaning_log_other_recode(x)
}
\arguments{
\item{x}{A list or data frame row representing the "other" response record.}
}
\value{
NULL. Updates the global variable \code{cleaning_log_other} in-place.
}
\description{
Routes a recoding action to the appropriate handler based on whether the referenced
question is a \code{select_one} or \code{select_multiple}.
}
