% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraR_arrange_with_spacing.R
\name{extraR_arrange_with_spacing}
\alias{extraR_arrange_with_spacing}
\title{Arrange With Spacing}
\usage{
extraR_arrange_with_spacing(df, col)
}
\arguments{
\item{df}{A data frame to be formatted}

\item{col}{A column to be used for grouping.}
}
\value{
A new data frame with with all operations performed.
}
\description{
This function takes a data frame and column then groups similar rows of that column
Arranges the Overall analysis into descending order and then add a blank space between groups
For easy readability of the analysis
}
