% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraR_create_table_variable_x_group.R
\name{extraR_create_table_variable_x_group}
\alias{extraR_create_table_variable_x_group}
\title{Create Table Variable X Group}
\usage{
extraR_create_table_variable_x_group(
  results_table,
  analysis_key = "analysis_key",
  value_columns = c("stat", "stat_low", "stat_upp"),
  extra_columns = NULL,
  list_for_excel = FALSE
)
}
\arguments{
\item{extra_columns}{All the columns that you want included in the output}

\item{xxx}{Use all params from create_table_variable_x_group}
}
\value{
A new data frame with with all operations performed.
}
\description{
This function extends the functionality of "create_table_variable_x_group" from analysis tools from
IMPACT HQ by providing an option of adding extra columns to the output
}
\examples{
# examples
extraR_create_table_variable_x_group(
  analysis_key = "label_analysis_key",
  value_columns = c("stat", "n", "n_total"),
  extra_columns = c("number", "sector", "indicator", "subset")
  )
}
