% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formated_output.R
\name{insert_empty_rows}
\alias{insert_empty_rows}
\title{Insert Empty Rows into a Data Frame}
\usage{
insert_empty_rows(df, col)
}
\arguments{
\item{df}{The data frame.}

\item{col}{The column used to identify groups.}
}
\value{
A modified dataset with empty rows separating distinct groups.
}
\description{
Internal helper function to insert empty rows between grouped elements.
}
\keyword{internal}
