% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_api.R
\name{kobo_export_poll}
\alias{kobo_export_poll}
\title{Poll Kobo Export Status}
\usage{
kobo_export_poll(export_url, sleep_time = 5, max_retries = 60)
}
\arguments{
\item{export_url}{The export URL returned by \code{kobo_export_trigger}}

\item{sleep_time}{Seconds to wait between polling}

\item{max_retries}{Maximum number of retries before timing out}
}
\value{
The download URL when successful
}
\description{
Poll Kobo Export Status
}
