% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_api.R
\name{kobo_export_trigger}
\alias{kobo_export_trigger}
\title{Trigger Kobo Export}
\usage{
kobo_export_trigger(
  asset_id,
  server_url = "https://kobo.impact-initiatives.org"
)
}
\arguments{
\item{asset_id}{The Kobo asset ID}

\item{server_url}{The Kobo server URL}
}
\value{
The export URL to poll
}
\description{
Trigger Kobo Export
}
