% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_api.R
\name{kobo_setup_token}
\alias{kobo_setup_token}
\title{Set Kobo API Token}
\usage{
kobo_setup_token(token = NULL)
}
\arguments{
\item{token}{Character. The Kobo API token. If NULL (default), the function
prompts for the token interactively.}
}
\description{
Securely saves the Kobo API token using keyring. If a token is not provided,
the user will be prompted to enter it interactively.
}
