% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_group_analysis.R
\name{my_create_table_variable_x_group}
\alias{my_create_table_variable_x_group}
\title{Create Table Variable x Group}
\usage{
my_create_table_variable_x_group(
  results_table,
  analysis_key = "analysis_key",
  value_columns = c("stat", "stat_low", "stat_upp"),
  extra_columns = NULL,
  list_for_excel = FALSE
)
}
\arguments{
\item{results_table}{the results}

\item{analysis_key}{the key}

\item{value_columns}{columns}

\item{extra_columns}{optional columns}

\item{list_for_excel}{internal use}
}
\value{
A pivotted dataframe
}
\description{
function to add sector and indicator to the table for output
}
