% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_other_responses.R
\name{save_other_responses}
\alias{save_other_responses}
\title{Save Other Responses}
\usage{
save_other_responses(
  df,
  ref_date = "",
  enumerator_id = NULL,
  save_location = "output",
  other_db = NULL
)
}
\arguments{
\item{df}{Data frame containing the responses to write.}

\item{ref_date}{Reference date for the filename (default is "").}

\item{enumerator_id}{Optional string indicating the enumerator id column (default is NULL).}

\item{save_location}{Directory to save the output file (default is "output").}

\item{other_db}{Data frame containing dropdown mapping for existing choices (default is NULL).}
}
\value{
NULL. Saves an Excel file.
}
\description{
This function saves the other responses into an Excel workbook with specific formatting and data validation.
The target columns are identified dynamically by name, allowing flexibility such as adding an \code{enumerator_id}.
}
