% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_api.R
\name{kobo_download_data}
\alias{kobo_download_data}
\title{Download Kobo Data}
\usage{
kobo_download_data(
  asset_id,
  server_url = "https://kobo.impact-initiatives.org",
  cache_dir = "data/",
  filename = paste0("kobo_data_", asset_id, ".xlsx")
)
}
\arguments{
\item{asset_id}{The Kobo asset ID}

\item{server_url}{The Kobo server URL (default: "https://kobo.impact-initiatives.org")}

\item{cache_dir}{Directory to save the downloaded file}

\item{filename}{The name of the downloaded file (default: "kobo_data_<asset_id>.xlsx")}
}
\value{
The path to the downloaded Excel file
}
\description{
Main function to download data from Kobo.
}
